ml/requirements.txt
# Python ML Dependencies for UPR
# Install with: pip install -r ml/requirements.txt

# Core ML libraries
xgboost>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.0

# Model persistence
joblib>=1.3.0

# Explainability
shap>=0.44.0

# Optional: Deep learning (for future enhancements)
# torch>=2.0.0
# transformers>=4.30.0

# Optional: Advanced NLP
# spacy>=3.6.0
# nltk>=3.8.0

# Optional: Time series forecasting
# prophet>=1.1.0
# statsmodels>=0.14.0
